[plugin-visual] Remove deprecated non-unified definitions of cut strategies for mobile apps
- Deprecated screenshot configuration parameter `nativeFooterToCut` is remove, use `cutBottom` parameter for mobile applications.
- Deprecated property `mobile.screenshot.strategy.<YOUR_STRATEGY_NAME>.native-footer-to-cut` is removed, use `mobile.screenshot.strategy.<YOUR_STRATEGY_NAME>.cut-bottom`
