--- vividus-plugin-mobile-app/src/main/resources/properties/deprecated/deprecated.properties
@@ -1,1 +1,0 @@
-mobile\.screenshot\.strategy\.([^.]+)\.native-footer-to-cut=mobile.screenshot.strategy.$1.cut-bottom
